+++ src/App.tsx
@@ -1,32 +1,12 @@
-import { Outlet, NavLink } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 import './App.scss';
-import classNames from 'classnames';
+import { Navbar } from './components/Navbar/Navbar';
 
-const getLinkClass = ({ isActive }: { isActive: boolean }) =>
-  classNames('navbar-item', {
-    'has-background-grey-lighter': isActive,
-  });
+
 
 export const App = () => (
   <div data-cy="app">
-    <nav
-      data-cy="nav"
-      className="navbar is-fixed-top has-shadow"
-      role="navigation"
-      aria-label="main navigation"
-    >
-      <div className="container">
-        <div className="navbar-brand">
-          <NavLink to="/" className={getLinkClass}>
-            Home
-          </NavLink>
-
-          <NavLink to="/people" className={getLinkClass}>
-            People
-          </NavLink>
-        </div>
-      </div>
-    </nav>
+    <Navbar />
 
     <main className="section">
       <div className="container">
