+++ src/Root.tsx
@@ -6,24 +6,21 @@ import {
 } from 'react-router-dom';
 
 import { App } from './App';
-import { PeopleProvider } from './contexts/PeopleContext';
 import { PageNotFound } from './components/PageNotFound/PageNotFound';
 import { HomePage } from './components/HomePage/HomePage';
 import { PeoplePage } from './components/People/PeoplePage';
 
 export const Root = () => (
   <Router>
-    <PeopleProvider>
-      <Routes>
-        <Route path="/" element={<App />}>
-          <Route index element={<HomePage />} />
-          <Route path="home" element={<Navigate to="/" replace />} />
-          <Route path="people">
-            <Route path=":slug?" element={<PeoplePage />}></Route>
-          </Route>
-          <Route path="*" element={<PageNotFound />} />
+    <Routes>
+      <Route path="/" element={<App />}>
+        <Route index element={<HomePage />} />
+        <Route path="home" element={<Navigate to="/" replace />} />
+        <Route path="people">
+          <Route path=":slug?" element={<PeoplePage />}></Route>
         </Route>
-      </Routes>
-    </PeopleProvider>
+        <Route path="*" element={<PageNotFound />} />
+      </Route>
+    </Routes>
   </Router>
 );
